pyproject.toml
[project]
name = "miniclaw"
version = "0.1.3"
description = "mini claw"
authors = [
    {name = "ztcz",email = "contact@ztenith.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "loguru (>=0.7.3,<0.8.0)",
    "shortuuid (>=1.0.13,<2.0.0)",
    "pydantic (>=2.13.4,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "aiofiles (>=25.1.0,<26.0.0)",
    "fastapi (>=0.136.1,<0.137.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "litellm (>=1.82.6,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "mss (>=10.2.0,<11.0.0)",
    "prompt-toolkit (>=3.0.52,<4.0.0)",
    "aioconsole (>=0.8.2,<0.9.0)"

]

[tool.poetry.group.dev.dependencies]
toml = "^0.10.2"
pytest-asyncio = "^1.3.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "nexus"
url = "https://mirrors.aliyun.com/pypi/simple/"
priority = "primary"



